consensus: Move IterationCtx struct to a separate file
