+++ consensus/src/proposal/handler.rs
@@ -11,7 +11,7 @@ use crate::step_votes_reg::SafeCertificateInfoRegistry;
 use crate::user::committee::Committee;
 use async_trait::async_trait;
 
-use crate::execution_ctx::RoundCommittees;
+use crate::iteration_ctx::RoundCommittees;
 use node_data::message::{Message, Payload};
 use std::sync::Arc;
 use tokio::sync::Mutex;
