+++ consensus/src/ratification/handler.rs
@@ -15,7 +15,7 @@ use tracing::{error, warn};
 use crate::aggregator::Aggregator;
 
 use crate::config;
-use crate::execution_ctx::RoundCommittees;
+use crate::iteration_ctx::RoundCommittees;
 use crate::quorum::verifiers::verify_votes;
 use node_data::message::payload::{QuorumType, Ratification, ValidationResult};
 use node_data::message::{payload, Message, Payload, Topics};
